Eval の実装を差し替え
